Add files via upload
# Extended Description

This repository is designed to help you learn and understand various aspects of Node.js through practical examples and projects. Each folder contains a specific topic or project that demonstrates different functionalities and features of Node.js. Below is an extended description of each section:

## ES Module

The ES Module folder provides examples of how to use ES6 modules in Node.js. It includes basic arithmetic functions and demonstrates how to import and use these functions in other files.

## Events

The Events folder contains examples of using the EventEmitter class in Node.js. It includes basic event handling and a challenge to handle multiple user events and track their occurrences. This is useful for understanding how to work with events in Node.js applications.

## fs_module

The fs_module folder contains examples of using the `fs` module for file system operations. It includes examples of asynchronous file operations using callbacks, promises, and async/await, as well as synchronous file operations. This is useful for understanding how to work with the file system in Node.js.

## HTTP Module

The HTTP Module folder contains an example of creating an HTTP server using Node.js. It includes a simple HTTP server that responds to different routes. This is useful for understanding how to create and work with HTTP servers in Node.js.

## Import and Export Module

The Import and Export Module folder demonstrates different ways of importing and exporting modules in Node.js. It includes examples of importing and using functions from other files, as well as using built-in Node.js modules like `os` and `path`. This is useful for understanding how to organize and structure your Node.js applications.

## Mega_Project 1

The Mega_Project 1 folder contains a larger project demonstrating a URL shortener application. It includes the main application file, a JSON file to store the shortened URLs, and front-end HTML and CSS files. This is useful for understanding how to build and structure larger Node.js applications.

## Mini_Project 1

The Mini_Project 1 folder contains a simple ToDo list application. It includes a basic ToDo list application, a more advanced ToDo list application with file storage, and a JSON file to store the tasks. This is useful for understanding how to build simple applications and work with file storage in Node.js.

## Mini_Project 2

The Mini_Project 2 folder contains a file system program. It includes a program to create, append, and read files, as well as a sample text file. This is useful for understanding how to work with the file system in Node.js.

## Mini_Project 3

The Mini_Project 3 folder contains a random joke generator. It includes examples of fetching a random joke from an API and displaying it. This is useful for understanding how to work with APIs in Node.js.

## Mini_Project 4

The Mini_Project 4 folder contains a currency converter application. It includes a program to convert USD to other currencies using an API. This is useful for understanding how to work with APIs and perform currency conversions in Node.js.

## Mini_Project 5

The Mini_Project 5 folder contains a weather information application. It includes a program to fetch and display weather information for a given city using an API. This is useful for understanding how to work with APIs and display weather information in Node.js.

## Getting Started

To run any of the projects or programs, navigate to the respective folder and run the following command: